Reorganized Run Flow
