+++ index.js
@@ -5,6 +5,7 @@ const chalk = require('chalk')
 
 // require scraping tools
 const scrape = require('./src/scraper')
+const analyze = require('./src/analyze')
 
 program
   .option('-u, --URL <s>', 'The URL to start scraping')
@@ -12,6 +13,7 @@ program
   .action(function(url, dir) {
     console.log(chalk.bgMagenta(`scraping ${chalk.green(url)} to ${chalk.green(dir)}`))
     scrape(url, dir)
+    analyze(url, dir)
   })
 
 program.parse(process.argv)
