+++ package.json
@@ -19,6 +19,8 @@
     "cheerio": "^1.0.0-rc.2",
     "commander": "^2.17.1",
     "crypto-random-string": "^1.0.0",
+    "klaw": "^3.0.0",
+    "mime": "^2.3.1",
     "mkdirp": "^0.5.1",
     "node-fetch": "^2.2.0",
     "string-similarity": "^1.2.1",
